+++ src/com/kallasoft/smugmug/api/json/v1_2_0/images/ChangeSettings.java
@@ -111,7 +111,7 @@ public class ChangeSettings extends AbstractMethod {
 	 * @see #execute(String, String[])
 	 */
 	public ChangeSettingsResponse execute(String url, String apiKey,
-			String sessionID, Integer imageID, Integer albumID) {
+			String sessionID, Long imageID, Long albumID) {
 		return execute(url, new String[] { apiKey, sessionID,
 				APIUtils.toString(imageID), APIUtils.toString(albumID) });
 	}
@@ -147,7 +147,7 @@ public class ChangeSettings extends AbstractMethod {
 	 * @see #execute(String, String[])
 	 */
 	public ChangeSettingsResponse execute(String url, String apiKey,
-			String sessionID, Integer imageID, String caption, String keywords,
+			String sessionID, Long imageID, String caption, String keywords,
 			Boolean hidden) {
 		return execute(url, new String[] { apiKey, sessionID,
 				APIUtils.toString(imageID), null, caption, keywords,
