+++ src/com/kallasoft/smugmug/api/json/v1_2_0/login/WithHash.java
@@ -131,7 +131,7 @@ public class WithHash extends AbstractMethod {
 	 * 
 	 * @see #execute(String, String[])
 	 */
-	public WithHashResponse execute(String url, String apiKey, Integer userID,
+	public WithHashResponse execute(String url, String apiKey, Long userID,
 			String passwordHash) {
 		return execute(url, new String[] { apiKey, APIUtils.toString(userID),
 				passwordHash });
