+++ src/com/kallasoft/smugmug/api/json/v1_2_0/subcategories/Rename.java
@@ -101,7 +101,7 @@ public class Rename extends AbstractMethod {
 	 * @see #execute(String, String[])
 	 */
 	public RenameResponse execute(String url, String apiKey, String sessionID,
-			Integer subCategoryID, String name) {
+			Long subCategoryID, String name) {
 		return execute(url, new String[] { apiKey, sessionID,
 				APIUtils.toString(subCategoryID), name });
 	}
