+++ src/com/kallasoft/smugmug/api/json/v1_2_1/images/ChangePosition.java
@@ -55,7 +55,7 @@ public class ChangePosition extends
 
 	@Override
 	public ChangePositionResponse execute(String url, String apiKey,
-			String sessionID, Integer imageID, Integer position) {
+			String sessionID, Long imageID, Integer position) {
 		return execute(url, new String[] { apiKey, sessionID,
 				APIUtils.toString(imageID), APIUtils.toString(position) });
 	}
