+++ src/com/kallasoft/smugmug/api/json/v1_2_1/login/WithHash.java
@@ -53,7 +53,7 @@ public class WithHash extends
 	}
 
 	@Override
-	public WithHashResponse execute(String url, String apiKey, Integer userID,
+	public WithHashResponse execute(String url, String apiKey, Long userID,
 			String passwordHash) {
 		return execute(url, new String[] { apiKey, APIUtils.toString(userID),
 				passwordHash });
